changed migations

+++ db/migrate/20121008184824_change_data_type_for_total_height.rb
@@ -8,8 +8,8 @@ class ChangeDataTypeForTotalHeight < ActiveRecord::Migration
 
   def down
   	change_table :population_states do |t|
-  		t.change :total_male_height, :float
-			t.change :total_female_height, :float
+  		t.change :total_male_height, :integer
+			t.change :total_female_height, :integer
   	end
   end
 end
